+++ stubs/amaranth/hdl/_ast.pyi
