+++ stubs/amaranth/hdl/_cd.pyi
