+++ stubs/amaranth/hdl/_dsl.pyi
