+++ stubs/amaranth/hdl/_ir.pyi
