+++ stubs/amaranth/hdl/_mem.pyi
