+++ stubs/amaranth/hdl/_xfrm.pyi
